==> notebooks/01_data_extraction.ipynb <==
# Data Extraction
---

# import libraries
import pandas as pd 
import kagglehub # import data via Kaggle API
import os

import warnings
warnings.filterwarnings("ignore")

# Download data from Kaggle API to local machine
path = kagglehub.dataset_download("nih-chest-xrays/data")

print("Path to dataset files:", path)

# extract data entry file
data_entry_df = pd.read_csv(os.path.join(path,'Data_Entry_2017.csv'))
data_entry_df.head()

# extract test split list
test_list = pd.read_csv(os.path.join(path, 'test_list.txt'))
test_list.shape

# extract train split list
train_val_list = pd.read_csv(os.path.join(path, 'train_val_list.txt'))
train_val_list.shape


# Save extracted data to the interim folder as a pickle file

data_entry_df.to_pickle('../data/interim/data_entry_df.pkl')
test_list.to_pickle('../data/interim/test_list.pkl')
train_val_list.to_pickle('../data/interim/train_val_list.pkl')
