==> notebooks/02_data_cleaning.ipynb <==
# Data Cleaning
---

# import libraries
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

# load extracted data

data_entry_df = pd.read_pickle('../data/interim/data_entry_df.pkl')
train_val_list = pd.read_pickle('../data/interim/train_val_list.pkl')
test_list = pd.read_pickle('../data/interim/test_list.pkl')

# removed unnamed column
data_entry_df = data_entry_df.drop(columns=["Unnamed: 11"],axis=1)

# rename columns 
data_entry_df.columns = data_entry_df.columns.tolist() 
data_entry_df.columns = ["image_index", "finding_labels", "follow_up_number", "patient_id", "patient_age", "patient_gender", "view_position", "original_img_width", "original_img_height", "img_pixel_spacing_x", "img_pixel_spacing_y"]
data_entry_df.head()

# checking basic characteristic of the dataset (na values, null values, duplicated values, and checked data type of each column)

print(f'Checking for NA values: \n {data_entry_df.isna().sum()}')
print(f'Checking for Null values: \n {data_entry_df.isnull().sum()}')
print(f'Checking for duplicated values: \n {data_entry_df.duplicated().sum()}')
print(f'Checking data type of each column: \n {data_entry_df.info()}')

# issue: missing column headers in both test_list and train_val_list datasets
# added a 'image_index' as the header for both datasets

shifted_train_val = train_val_list.shift(1)
shifted_train_val.columns = ['image_index']
shifted_train_val.iloc[0,0] = train_val_list.columns[0]
print(shifted_train_val.head())

shifted_test_list = test_list.shift(1)
shifted_test_list.columns = ['image_index']
shifted_test_list.iloc[0,0] = test_list.columns[0]
print(shifted_test_list.head())

# the data_entry dataframe is structured per image. As a result, there are multiple x-ray images that belong to the same patient from follow-ups
# created a seperate patient-level dataset for EDA 

patient_data = data_entry_df.groupby(["patient_id", "follow_up_number"]).agg({
    "image_index": "first",
    "finding_labels": "first",
    "patient_age": "first",
    "patient_gender": "first",
    "view_position": "first"
})
patient_data.head(15)

# Save to .pkl files

data_entry_df.to_pickle('../data/interim/data_entry_df.pkl')
patient_data.to_pickle('../data/interim/patient_data.pkl')
shifted_train_val.to_pickle('../data/interim/train_val_list.pkl')
shifted_test_list.to_pickle('../data/interim/test_list.pkl')
